Client GUI can access all functionality, next GUI design

+++ PhilosopherStoneAndroidClient/app/src/main/java/com/sisteritb/philosopherstone/connection/request/MixItemRequest.java
@@ -17,8 +17,8 @@ import org.json.simple.JSONValue;
 public class MixItemRequest extends Request{
     
     public String token;
-    public int item1;
-    public int item2;
+    public long item1;
+    public long item2;
 
     public MixItemRequest() {
         method = "mixitem";
